fix: 修改关于Raspberry Pi OS(Debian Trixie)下串口名称及DAC输出电流映射范围 调整厚度数据转换电流计算公式，防止DAC输出电流超出4~20mA范围

+++ dac8760.h
@@ -20,7 +20,7 @@ public:
     void DAC8760_Write(uint8_t Addr, uint16_t Data);
     uint32_t DAC8760_Read(uint8_t Addr);
     void Delay(uint32_t nCount);
-    void elec_write(float value);
+    void elec_write(double value);
 
 public slots:
     void send_thick(double thickness);
